Remove some CF-ism
Removed the CF-specific fields and started defining a profile doc
for the new 'context' object.

Changes per https://github.com/openservicebrokerapi/servicebroker/issues/115#issuecomment-283080852

Make "context" required on instance update

while in there I added the word "instance" to this sentence:
`Platform specific contextual information under which the service instance is to be provisioned.`

To make it clear we're provisioning a service instance and not a service.

Signed-off-by: Doug Davis <dug@us.ibm.com>