Combine all docs into one
This is the first step. Having everything in one file reduces the
friction of dealing with individual files. This makes it easier to start
moving and deleting sections from a global perspective.

This is almost a pure copy. There was some rearranging to bring to the
top sections that will probably remain.

Additions of a reference to doc formatting rfc.

Some tables were converted from html to markdown as they were not
rendering correctly.

Signed-off-by: Morgan Bauer <mbauer@us.ibm.com>