Make sure strings can't be empty when appropriate
I decided that while in most cases an empty string can be treated the
same as the property being absent, when it came to IDs I thought it was
best to make it clear that if an optional ID is present then it MUST
have a value. Otherwise someone could write code to deal with an empty
string ID and that might have different semantics from no ID at all.
E.g. construct a string of the form xxx?id=  when there is a value
(even if empty), but if its missing it might construct it as just "xxx".
And having that difference would be an interop problem.

Signed-off-by: Doug Davis <dug@us.ibm.com>