Clarify how to clean-up during a failed provision/bind
Closes #348

We actually already had text to deal with this in the bind() case, it
was just missing the MUST, so I added that and then copied that same
text into the provision case.

Net: platforms MUST send a DELETE to the brokers upon a failed
provison/bind just to make sure things are cleaned up.

I'll open a 2nd PR to add text talking about how the brokers SHOULD
clean-up on their own w/o the need for the DELETE.

Signed-off-by: Doug Davis <dug@us.ibm.com>