Add 'description' field to the 'maintenance_info.'
This tackles the `There SHOULD be some indication of the impact of an update`
part of #641.

This is not necessarily about a description string containing versions changes
only, this may contain details about configuration changes, the default value
changes, etc.

The main use case for `maintenance_info` and `description` field is for the
plans that need maintenance done to them every 2-4 weeks or so, thus creating a
new plan every time is not an option.

[#165560405](https://www.pivotaltracker.com/story/show/165560405)