Clarify how maintenance_info objects should be compared.
- MaintenanceInfoConflict error should only be produced when
`maintenance_info.version` doesn't match. Other fields of Maintenance
Info object should be ignored in such checks.
- Adjust wording in the request/response body definitions to reflect the
above better.

This change is needed for the following reasons:
- The typo fix in the `description` of the Maintenance Info object
should not cause new `maintenance_info` version to be created.
- Platforms don't have to store or send back to the Service Broker all
the fields of the `maintenance_info`.
- In general, the `maintenance_info.version` is what matters from
perspective of checking whether two objects are same or not.

[#165995791](https://www.pivotaltracker.com/story/show/165995791)