Clarify version of semantic versioning of maintenance_info.version
So that when semver 2.1 or 3.0 gets released, it would be clear which
version should Brokers and Platforms implement, and will allow OSBAPI
group to update the referenced version in one of the next OSBAPI
releases when it makes sense to do so. It gives more control over what
that version field even means.

Co-authored-by: Niki Maslarski <nikolay.maslarski@sap.com>