Remove ambiguity on provisioning status code
The specification for synchronous provisioning says `201 Created` status code
should be returned when for provisioning and bind and `200 OK` only for
update, unbind and deprovisioning requests.

This also reverts #561 and assumes duplicated requests will return `409
Conflict` for the same unique UUID and it should be advised to platforms
to start polling the `last_operation` endpoint to wait for completion.